pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robot_framework"
version = "0.0.1"
authors = [
  { name="ITK Development", email="itk-rpa@mkb.aarhus.dk" },
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
]
dependencies = [
    "OpenOrchestrator == 1.*",
    "Pillow == 9.5.0",
    "docx-mailmerge2 == 0.8.0",
    "docx2pdf == 0.1.8",
    "python-serviceplatformen == 1.*",
]

[project.optional-dependencies]
dev = [
  "pylint",
  "flake8"
]